Split up DevicePickerViewController
