+++ Client/Frontend/DevicePickerTableViewCell.swift
@@ -1,0 +1,55 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/
+
+import UIKit
+
+class DevicePickerTableViewCell: UITableViewCell {
+
+    private struct UX {
+        static let deviceRowTextFont = UIFont.systemFont(ofSize: 16)
+        static let deviceRowTextPaddingLeft: CGFloat = 72
+        static let deviceRowTextPaddingRight: CGFloat = 50
+    }
+
+    static let CellIdentifier = "ClientPickerTableViewCell"
+
+    var nameLabel: UILabel
+    var checked: Bool = false {
+        didSet {
+            self.accessoryType = checked ? .checkmark : .none
+        }
+    }
+
+    var clientType = ClientType.Mobile {
+        didSet {
+            self.imageView?.image = UIImage.templateImageNamed(clientType.rawValue)
+        }
+    }
+
+    override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
+        nameLabel = UILabel()
+        super.init(style: style, reuseIdentifier: reuseIdentifier)
+        contentView.addSubview(nameLabel)
+        nameLabel.font = UX.deviceRowTextFont
+        nameLabel.numberOfLines = 2
+        nameLabel.lineBreakMode = .byWordWrapping
+        self.tintColor = UIColor.label
+        self.preservesSuperviewLayoutMargins = false
+        self.selectionStyle = .none
+    }
+
+    override func layoutSubviews() {
+        super.layoutSubviews()
+
+        nameLabel.snp.makeConstraints { (make) -> Void in
+            make.left.equalTo(UX.deviceRowTextPaddingLeft)
+            make.centerY.equalTo(self.snp.centerY)
+            make.right.equalTo(self.snp.right).offset(-UX.deviceRowTextPaddingRight)
+        }
+    }
+
+    required init(coder aDecoder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+}
